Started From Scratch by using the Drag Drop Lab
This gives us layout, drag/drop capabilities...  I put the Context Menu
in it, but it might not work.
Open issues:
Dragging existing item causes a new item to be created
Not sure if can rotate/scale the items
If drag item on top of another, causes first to jump location
ListBox orientation on the sides and general layout not "pretty"
Need to separate the lists so they can have different ordered content
Need to check the clear data functionality - not sure if curClipArt list
is correct
Need to add the "join" button in the grid locations and have a Button
for onclick to replace the content with the listboxes already in this
version.